added equipment entities, views and controllers

+++ src/main/java/com/prod/hydraulicsystemsmaintenance/entities/Technician.java
@@ -1,15 +1,15 @@
 package com.prod.hydraulicsystemsmaintenance.entities;
 
 public class Technician extends User {
-    public Technician(int id, String name, String username, String password, int administrator) {
+    public Technician(Long id, String name, String username, String password, Integer administrator) {
         super(id, name, username, password, administrator);
     }
 
-    public Technician(String name, String username, String password, int administrator) {
+    public Technician(String name, String username, String password, Integer administrator) {
         super(name, username, password, administrator);
     }
 
-    public Technician(Integer id, String name, String username, Integer administrator) {
+    public Technician(Long id, String name, String username, Integer administrator) {
         super(id, name, username, administrator);
     }
 }
